pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "Multi-agent icon generator using pydantic-ai delegation pattern"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "pydantic-ai>=1.6.0",
    "httpx>=0.27.0",
]
